requirements.txt
requests
python-dotenv
geopy
python-socketio
eventlet
websocket-client